pyproject.toml
[project]
name = "xnatio"
version = "0.2.0"
description = "Production-grade CLI for XNAT neuroimaging server administration"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Ricky Wong", email = "rickywonglt15@outlook.com"},
]
maintainers = [
    {name = "Ricky Wong", email = "rickywonglt15@outlook.com"},
]
license = {text = "MIT"}
keywords = ["xnat", "dicom", "neuroimaging", "cli", "admin"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: System :: Archiving",
    "Topic :: System :: Systems Administration",
    "Typing :: Typed",
]
dependencies = [
    "pyxnat>=1.6.3,<2.0",
    "python-dotenv>=1.0.1,<2.0",
    "requests>=2.31.0,<3.0",
    "pydicom>=2.4.4,<3.0",
    "tqdm>=4.67.1,<5.0",
    "pynetdicom>=2.0.2,<3.0",
]

[project.scripts]
xnatio = "xnatio.cli:run_cli"
xio = "xnatio.cli:run_cli"

[project.optional-dependencies]
dev = [
    "pre-commit>=3.7.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.6.9",
    "mypy>=1.8.0",
    "types-requests>=2.31.0",
    "pip-audit>=2.7.0",
]

[project.urls]
homepage = "https://github.com/rickyltwong/xnatio"
repository = "https://github.com/rickyltwong/xnatio.git"

[tool.uv]
package = true

# Build system configuration for packaging
[build-system]
requires = [
    "setuptools>=68",
    "wheel",
]
build-backend = "setuptools.build_meta"

[tool.setuptools]
zip-safe = false

[tool.setuptools.packages.find]
where = ["."]
include = ["xnatio*"]

[tool.ruff]
line-length = 100
target-version = "py38"
exclude = [
    "docs/legacy",
    "scripts",
    "xnatio.egg-info",
    "build",
    "dist",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["data"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
ignore_missing_imports = true
show_error_codes = true
pretty = true

[[tool.mypy.overrides]]
module = "pyxnat.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pynetdicom.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pydicom.*"
ignore_missing_imports = true

[tool.coverage.run]
source = ["xnatio"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
